Scraper rewrite
This update reflects a major revision to the scraper. The dataset increases by about 13% without the addition of a new season. The bulk of the increase is in Seasons 1 through 12.

Several other minor changes include updated column labels, the way NaN values are recorded, and a separation of clue value from Daily Double wager amount.

I will update with Season 39 data after the finale in late July.